Backend FINALIZADO

+++ backend/controllers/tareasController.js
@@ -2,7 +2,7 @@ const asyncHandler = require('express-async-handler')
 const Tarea = require('../model/tareasModel')
 
 const getTareas = asyncHandler(async (req, res) => {
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({ user: req.user.id })
 
     res.status(200).json(tareas)
 })
@@ -14,7 +14,8 @@ const setTareas = asyncHandler(async (req, res) => {
     }
 
     const tarea = await Tarea.create({
-        texto: req.body.texto
+        texto: req.body.texto,
+        user: req.user.id
     })
 
     res.status(201).json({ tarea })
@@ -22,28 +23,47 @@ const setTareas = asyncHandler(async (req, res) => {
 
 const updateTareas = asyncHandler(async (req, res) => {
 
+    //verificamos que la tarea existe
     const tarea = await Tarea.findById(req.params.id)
     if (!tarea) {
         res.status(400)
         throw new Error('La tarea no fué encontrada')
     }
-    const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
 
-    res.status(200).json(tareaUpdated)
+    //verificar que la tarea pertenezca al usuario del token que la quiere modificar
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Acceso no autorizado')
+    } else {
+        const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
+
+        res.status(200).json(tareaUpdated)
+    }
 })
 
 const deleteTareas = asyncHandler(async (req, res) => {
 
+    //verificamos que la tarea existe
     const tarea = await Tarea.findById(req.params.id)
     if (!tarea) {
         res.status(400)
         throw new Error('La tarea no fué encontrada')
     }
 
-    await Tarea.deleteOne(tarea)
-    //await Tarea.findByIdAndDelete(req.params.id)
+    //verificar que la tarea pertenezca al usuario del token que la quiere modificar
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Acceso no autorizado')
+    } else {
 
-    res.status(200).json({ id: req.params.id })
+        await Tarea.deleteOne(tarea)
+        //await Tarea.findByIdAndDelete(req.params.id)
+
+        res.status(200).json({ id: req.params.id })
+
+    }
+
+
 })
 
 module.exports = {
